requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cors==1.0.1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
jmespath==0.10.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tldextract==3.2.0
urllib3==1.26.8
zope.event==4.5.0
zope.interface==5.4.0
